Corrección de errores
se corrigio los errores de los forms para buscar, eliminar y modificar iva

+++ capaEntidades/claseIva.cs
@@ -9,6 +9,6 @@ namespace capaEntidades
     public class claseIva
     {
         public int id_iva { get; set; }
-        public double valor_iva { get; set; }
+        public decimal valor_iva { get; set; }
     }
 }
